Add `User` and `Group`; handle _f...annotations.
When a class references itself in a type hint, one needs to use
`__future__.annotations`. `Process.parent` is of type `Process`, so such
references are needed. This change results in the values in
`__annotations__' to become strings, and thus I need to use `eval` to
evaluate the string to a type object.